Move tests to different files
